user_service/requirements.txt
Django>=5.2.1
psycopg2-binary
python-dotenv
djangorestframework>=3.15.0
drf-yasg>=1.21.7
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
gunicorn 